requirements.txt
langchain
python-dotenv
PyPDF2
openai
numpy
streamlit
scikit-learn
langchain-openai